Fixing bad type in vector2 and 3

+++ includes/math/spk_vector2.hpp
@@ -79,8 +79,8 @@ namespace spk
 			}
 			else
 			{
-				x = p_object[L"x"].as<int>();
-				y = p_object[L"y"].as<int>();
+				x = p_object[L"x"].as<long>();
+				y = p_object[L"y"].as<long>();
 			}
 		}
